% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.orthology.r
\name{analyse.orthology}
\alias{analyse.orthology}
\title{Analyse orthology between two species}
\usage{
analyse.orthology(
  species1 = "mouse",
  species2 = "human",
  allHomologs = load.homologs()
)
}
\arguments{
\item{species1}{String with the name of the first species, i.e. "mouse"}

\item{species2}{String with the name of the second species, i.e. "human"}

\item{allHomologs}{Output of load.homologs() function}
}
\value{
List. The most important thing is orthologs_one2one.
}
\description{
Returns a table of 1:1 orthologs along with gene sets (in Homologene ID format) detailing how genes differ between the two species
}
\examples{
ortholog_data = analyse.orthology("human","mouse")
}
