% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.species.r
\name{check.species}
\alias{check.species}
\title{Checks that the species names, i.e. "mouse" can be linked 1:1 to a species name from the MGI homolog table}
\usage{
check.species(requestedSpecies, allSpecies)
}
\arguments{
\item{requestedSpecies}{String with the name of the first species, i.e. "mouse"}

\item{allSpecies}{String with the names of all species from the MGI homolog table}
}
\value{
The name of the species in the MGI homolog table.
}
\description{
Returns the linked species name from the MGI homolog table, i.e. if you pass "mouse" it returns "mouse, laboratory".
Essentially it's a convenience function so you don't need to write out "mouse, laboratory". It's intended primarily
for use within the packages other functions and should not be generally called seperately.
}
\examples{
allHomologs = load.homologs()
species1 = check.species(requestedSpecies="mouse",allSpecies=allHomologs$`Common Organism Name`)

}
