% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.symbol.from.homologID.r
\name{get.symbol.from.homologID}
\alias{get.symbol.from.homologID}
\title{Get gene symbol from homologID}
\usage{
get.symbol.from.homologID(
  homoIDs,
  species1,
  species2,
  ortholog_data,
  symbol_species
)
}
\arguments{
\item{homoIDs}{The list of homolog IDs to be converted to gene symbols}

\item{species1}{String with the name of the first species, i.e. "mouse"}

\item{species2}{String with the name of the second species, i.e. "human"}

\item{ortholog_data}{Output of analyse.orthology() function}

\item{symbol_species}{The name of the species for which gene symbols should be returned. I.e. "mouse"}
}
\value{
List. The gene symbols equivilenet to homolog IDs in homoIDs
}
\description{
Most outputs of this package are in the format of homolog IDs. To be useful they need to be converted to gene symbols. This functon does that.
}
\examples{
allHomologs = load.homologs()
ortholog_data = analyse.orthology("human","mouse",allHomologs)
homoIDs_bg     = ortholog_data$species2_allGenes
symbols_bg = get.symbol.from.homologID(homoIDs_bg,"human","mouse",ortholog_data,"mouse")

}
