% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.homologs.r
\name{load.homologs}
\alias{load.homologs}
\title{Load the MGI homolog table}
\usage{
load.homologs()
}
\value{
The full MGI homolog table
}
\description{
Downloads and formats the MGI homolog table
}
\examples{
allHomologs = load.homologs()
}
