% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ortholog_data_Mouse_Human}
\alias{ortholog_data_Mouse_Human}
\title{Table of 1:1 Human-->Mouse orthologs}
\format{
An object of class \code{list} of length 10.
}
\source{
The code to prepare the .Rda file file from the marker file is:
\code{
allHomologs = load.homologs()
ortholog_data_Mouse_Human = analyse.orthology("human","mouse",allHomologs)
save(ortholog_data_Mouse_Human,file="One2One/data/ortholog_data_Mouse_Human.rda")
}
}
\usage{
ortholog_data_Mouse_Human
}
\description{
A dataset containing the MGI and HGNC symbols. This table was generated using this package.
}
\keyword{datasets}
